Isolate matrix functions.
